Rename types.py to datatypes.py
